agents/requirements.txt
agno>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
sse-starlette>=2.0.0
hyperliquid-python-sdk>=0.5.0
pydantic>=2.0.0
